refactor(apiserver): move Kafka sender/receiver into secondary/primary
Per review: the Kafka transport has a clear direction, so it should not sit
in the shared `common` module.

- Kafka sender (ServerEventSenderPort) -> secondary (outbound)
- Kafka receiver (ServerEventReceiverPort) -> primary (inbound)

`common` now only keeps the genuinely shared piece: the in-memory event hub,
a single instance that backs both ports in standalone mode (so an event sent
on one side is observed on the other), plus the shared
UnsupportedEventProtocolError. Each direction's provider switches on the
configured protocol and falls back to the shared hub for in-memory mode.
